+++ HISTORY.rst
@@ -10,6 +10,7 @@ Historia zmian
 * lepsze matchowanie źródeł z PBN (#1064),
 * weryfikuj obecnośc ORCID w PBN dla niezmatchowanych autorów (#1054),
 * pobieraj wszystkie osoby z PBNu (b/n),
+* pole dla wpisania wartości, czy praca występuje w profilu ORCID autora (#1054),
 
 202105.64
 ---------